fix: bottomNavItem demend.kt로 이동
